Fix stuck command acknowledgement issue.
During a deployment, the server can respond to a command acknowledgement
with a 'Failed' or 'Succeeded' command status even though the command is
in progress. This causes the command poller to ignore the command and
begin polling for a new command. The result is that the deployment hangs
until the server (60-70 min.) timeout has elapsed.

This fix simply unsticks the deployment by echoing back the command
status with a `PutHostCommandComplete` request to the server.